requirements.txt
numpy
pandas
scikit-learn
pdfplumber
pytesseract
Pillow
pymupdf4llm